+++ tests/resources/keeper_helpers.py
@@ -1,87 +1,108 @@
 import json
 import logging
-import os
-import pathlib
+import uuid
 
-from ocean_utils.agreements.service_factory import ServiceDescriptor, ServiceFactory
-from ocean_utils.ddo.metadata import MetadataBase
-from ocean_utils.ddo.public_key_rsa import PUBLIC_KEY_TYPE_RSA
-from secret_store_client.client import Client as SecretStore
+from eth_utils import remove_0x_prefix
 from ocean_utils.agreements.service_agreement import ServiceAgreement
+from ocean_utils.agreements.service_factory import ServiceDescriptor, ServiceFactory
+from ocean_utils.agreements.service_types import ServiceTypes
+from ocean_utils.aquarius.aquarius import Aquarius
 from ocean_utils.ddo.ddo import DDO
-from ocean_utils.did import DID, did_to_id
+from ocean_utils.ddo.metadata import MetadataMain
+from ocean_utils.ddo.public_key_rsa import PUBLIC_KEY_TYPE_RSA
+from ocean_utils.did import DID, did_to_id, did_to_id_bytes
+from ocean_utils.utils.utilities import checksum
+from secret_store_client.client import Client as SecretStore
 
-from ocean_events_handler.util import keeper_instance, web3, get_config
-from tests.resources.mocks.aquarius import AquariusMock
-
-
-def new_ddo(metadata, account, keeper):
-    did = DID.did()
-    ddo_service_endpoint = AquariusMock.get_service_endpoint(did)
-    _ddo = DDO(did)
-    _ddo.add_public_key(did, account.address)
-    _ddo.add_authentication(did, PUBLIC_KEY_TYPE_RSA)
-
-    metadata['base']['checksum'] = _ddo.generate_checksum(did, metadata)
-    checksum = metadata['base']['checksum']
-    _ddo.add_proof(checksum, account, keeper.sign_hash(checksum, account))
-
-    encrypted_files = secret_store_encrypt(did, json.dumps(metadata['base']['files']), account)
-    index = 0
-    for file in metadata['base']['files']:
-        file['index'] = index
-        index = index + 1
-        del file['url']
-    metadata['base']['encryptedFiles'] = encrypted_files
-
-    service_descriptors = [
-        ServiceDescriptor.metadata_service_descriptor(
-            metadata, ddo_service_endpoint),
-        ServiceDescriptor.authorization_service_descriptor(
-            get_config().secret_store_url),
-        ServiceDescriptor.access_service_descriptor(
-            metadata[MetadataBase.KEY]['price'],
-            'http://localhost:3030',
-            'http://localhost:3030',
-            3600,
-            keeper.escrow_access_secretstore_template.address,
-            keeper.escrow_reward_condition.address)
-    ]
-    for service in ServiceFactory.build_services(did, service_descriptors):
-        _ddo.add_service(service)
-
-    return _ddo
+from ocean_events_handler.util import get_config, keeper_instance, web3
+from tests.conftest import get_sample_ddo
 
 
 def get_registered_ddo(account, providers=None):
     keeper = keeper_instance()
-    # _ddo = new_ddo(Metadata.get_example(), account, keeper)
-    base = os.path.realpath(__file__).split(os.path.sep)[1:-1]
-    path = pathlib.Path(os.path.join(os.path.sep, *base, 'data', 'ddo_sample.json'))
-    with open(path, 'r') as file_handle:
-        json_text = file_handle.read()
+    aqua = Aquarius('http://localhost:5000')
 
-    did = DID.did()
-    replacement_dict = {
-        'owner_address': account.address,
-        'did_id': did_to_id(did),
-        'agreement_template_id': keeper.escrow_access_secretstore_template.address
+    for did in aqua.list_assets():
+        aqua.retire_asset_ddo(did)
+
+    metadata = get_sample_ddo()['service'][0]['attributes']
+    metadata['main']['files'][0]['checksum'] = str(uuid.uuid4())
+    ddo = DDO()
+    ddo_service_endpoint = aqua.get_service_endpoint()
+
+    metadata_service_desc = ServiceDescriptor.metadata_service_descriptor(metadata,
+                                                                          ddo_service_endpoint)
+
+    access_service_attributes = {
+        "main": {
+            "name": "dataAssetAccessServiceAgreement",
+            "creator": account.address,
+            "price": metadata[MetadataMain.KEY]['price'],
+            "timeout": 3600,
+            "datePublished": metadata[MetadataMain.KEY]['dateCreated']
+        }
     }
-    json_text = json_text % replacement_dict
-    _ddo = DDO(json_text=json_text)
-    ddo_service_endpoint = AquariusMock.get_service_endpoint(did)
 
-    text_for_sha3 = _ddo.metadata['base']['checksum'][2:]
-    keeper.did_registry.register(
-        _ddo.asset_id,
-        checksum=web3().sha3(text=f'{text_for_sha3}'),
+    service_descriptors = [ServiceDescriptor.authorization_service_descriptor(
+        'http://localhost:12001')]
+    service_descriptors += [ServiceDescriptor.access_service_descriptor(
+        access_service_attributes,
+        'http://localhost:8030',
+        keeper.escrow_access_secretstore_template.address
+    )]
+
+    service_descriptors = [metadata_service_desc] + service_descriptors
+
+    services = ServiceFactory.build_services(service_descriptors)
+    checksums = dict()
+    for service in services:
+        checksums[str(service.index)] = checksum(service.main)
+
+    # Adding proof to the ddo.
+    ddo.add_proof(checksums, account)
+
+    did = ddo.assign_did(DID.did(ddo.proof['checksum']))
+
+    stype_to_service = {s.type: s for s in services}
+    access_service = stype_to_service[ServiceTypes.ASSET_ACCESS]
+
+    name_to_address = {cname: cinst.address for cname, cinst in keeper.contract_name_to_instance.items()}
+    access_service.init_conditions_values(did, contract_name_to_address=name_to_address)
+    ddo.add_service(access_service)
+    for service in services:
+        ddo.add_service(service)
+
+    ddo.proof['signatureValue'] = keeper.sign_hash(did_to_id_bytes(did), account)
+
+    ddo.add_public_key(did, account.address)
+
+    ddo.add_authentication(did, PUBLIC_KEY_TYPE_RSA)
+
+    encrypted_files = do_secret_store_encrypt(
+        remove_0x_prefix(ddo.asset_id),
+        json.dumps(metadata['main']['files']),
+        account,
+        get_config()
+    )
+    _files = metadata['main']['files']
+    # only assign if the encryption worked
+    if encrypted_files:
+        index = 0
+        for file in metadata['main']['files']:
+            file['index'] = index
+            index = index + 1
+            del file['url']
+        metadata['encryptedFiles'] = encrypted_files
+
+    keeper_instance().did_registry.register(
+        ddo.asset_id,
+        checksum=web3().toBytes(hexstr=ddo.asset_id),
         url=ddo_service_endpoint,
         account=account,
         providers=providers
     )
-    AquariusMock.publish_asset_ddo(_ddo)
-
-    return _ddo
+    aqua.publish_asset_ddo(ddo)
+    return ddo
 
 
 def get_secret_store_client(account, config=None):
@@ -92,6 +113,17 @@ def get_secret_store_client(account, config=None):
         account.address,
         account.password
     )
+
+
+def do_secret_store_encrypt(did_id, document, provider_acc, config):
+    secret_store = SecretStore(
+        config.secret_store_url,
+        config.parity_url,
+        provider_acc.address,
+        provider_acc.password
+    )
+    encrypted_document = secret_store.publish_document(did_id, document)
+    return encrypted_document
 
 
 def secret_store_encrypt(did, document, account):
@@ -108,12 +140,12 @@ def secret_store_decrypt(did, encrypted_document, account):
     )
 
 
-def place_order(publisher_account, service_definition_id, ddo, consumer_account):
+def place_order(publisher_account, ddo, consumer_account):
     keeper = keeper_instance()
     agreement_id = ServiceAgreement.create_new_agreement_id()
     agreement_template = keeper.escrow_access_secretstore_template
     publisher_address = publisher_account.address
-    service_agreement = ServiceAgreement.from_ddo(service_definition_id, ddo)
+    service_agreement = ServiceAgreement.from_ddo(ServiceTypes.ASSET_ACCESS, ddo)
     condition_ids = service_agreement.generate_agreement_condition_ids(
         agreement_id, ddo.asset_id, consumer_account.address, publisher_address, keeper)
     time_locks = service_agreement.conditions_timelocks